requirements.txt
python-telegram-bot
python-dotenv
sqlalchemy
sqlmodel
asyncpg
httpx
